+++ packages/f-checkout/CHANGELOG.md
@@ -3,6 +3,20 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.4.0
+------------------------------
+*October 22, 2020*
+
+### Added
+- Address component to be displayed if `checkoutMethod` is set to 'Delivery'.
+
+### Removed
+- Address related form fields from `f-checkout`.
+
+### Changed
+- Checkout Component to replicate Collection and Delivery in ConsumerWeb.
+- Selector to uses `method` prop to change label text.
+
 
 v0.3.0
 ------------------------------